Include a guide for setting up Python
The usual instructions of "just download Anaconda" don't seem to
be working very well. Besides being a large download, `conda` is
just too slow and it's been almost impossible to update the packages
and install the latest Fatiando tools. The updated instructions use
Mambaforge so we get a small download with minimal dependencies
and `mamba` (a much faster version of `conda`).
For now, only includes Windows instructions.